# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyOrbit API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  name: string;
  subject: string;
  studyTime: string;
  skillLevel: string;
  goalType: string;
  skillCategories: string[];
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface CreateUserBody {
  name: string;
  subject: string;
  studyTime: string;
  skillLevel: string;
  goalType: string;
  skillCategories: string[];
  /** @nullable */
  bio?: string | null;
}

export interface UpdateUserBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  studyTime?: string | null;
  /** @nullable */
  skillLevel?: string | null;
  /** @nullable */
  goalType?: string | null;
  /** @nullable */
  skillCategories?: string[] | null;
  /** @nullable */
  bio?: string | null;
}

export interface MatchBreakdown {
  subjectMatch: number;
  studyTimeMatch: number;
  skillCategoryMatch: number;
  skillLevelMatch: number;
  goalTypeMatch: number;
}

export interface UserMatch {
  user: User;
  score: number;
  breakdown: MatchBreakdown;
}

export interface RoleRequirement {
  id: number;
  skillCategory: string;
  count: number;
  filled: number;
}

export interface TeamMember {
  id: number;
  userId: number;
  teamId: number;
  role: string;
  user: User;
  joinedAt: string;
}

export interface Team {
  id: number;
  name: string;
  subject: string;
  studyTime: string;
  goalType: string;
  minSkillLevel: string;
  requiredRoles: RoleRequirement[];
  members: TeamMember[];
  createdAt: string;
}

export interface CreateRoleRequirement {
  skillCategory: string;
  count: number;
}

export interface CreateTeamBody {
  name: string;
  subject: string;
  studyTime: string;
  goalType: string;
  minSkillLevel: string;
  requiredRoles: CreateRoleRequirement[];
}

export interface AddTeamMemberBody {
  userId: number;
  role: string;
}

export interface SubjectCount {
  subject: string;
  count: number;
}

export interface SkillLevelCount {
  level: string;
  count: number;
}

export interface StudyTimeCount {
  time: string;
  count: number;
}

export interface DashboardStats {
  totalUsers: number;
  totalTeams: number;
  bySubject: SubjectCount[];
  bySkillLevel: SkillLevelCount[];
  byStudyTime: StudyTimeCount[];
}

export type ListUsersParams = {
  /**
   * @nullable
   */
  subject?: string | null;
  /**
   * @nullable
   */
  skillCategory?: string | null;
  /**
   * @nullable
   */
  studyTime?: string | null;
  /**
   * @nullable
   */
  skillLevel?: string | null;
};

export type GetUserMatchesParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetTeamCandidatesParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};