# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyOrbit API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User profile operations
  - name: teams
    description: Team operations
  - name: matching
    description: Matching and discovery operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: subject
          in: query
          schema:
            type: ["string", "null"]
        - name: skillCategory
          in: query
          schema:
            type: ["string", "null"]
        - name: studyTime
          in: query
          schema:
            type: ["string", "null"]
        - name: skillLevel
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create or update a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}/matches:
    get:
      operationId: getUserMatches
      tags: [matching]
      summary: Get top matched users for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top matches with compatibility scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserMatch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"

  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /teams/{id}/candidates:
    get:
      operationId: getTeamCandidates
      tags: [matching]
      summary: Get top candidate users for a team role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top candidates for team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserMatch"

  /teams/{id}/members:
    post:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a member to a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeamMemberBody"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [matching]
      summary: Get galaxy dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        studyTime:
          type: string
        skillLevel:
          type: string
        goalType:
          type: string
        skillCategories:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - subject
        - studyTime
        - skillLevel
        - goalType
        - skillCategories
        - createdAt

    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        studyTime:
          type: string
        skillLevel:
          type: string
        goalType:
          type: string
        skillCategories:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
      required:
        - name
        - subject
        - studyTime
        - skillLevel
        - goalType
        - skillCategories

    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        studyTime:
          type: ["string", "null"]
        skillLevel:
          type: ["string", "null"]
        goalType:
          type: ["string", "null"]
        skillCategories:
          type: ["array", "null"]
          items:
            type: string
        bio:
          type: ["string", "null"]

    UserMatch:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        score:
          type: integer
        breakdown:
          $ref: "#/components/schemas/MatchBreakdown"
      required:
        - user
        - score
        - breakdown

    MatchBreakdown:
      type: object
      properties:
        subjectMatch:
          type: integer
        studyTimeMatch:
          type: integer
        skillCategoryMatch:
          type: integer
        skillLevelMatch:
          type: integer
        goalTypeMatch:
          type: integer
      required:
        - subjectMatch
        - studyTimeMatch
        - skillCategoryMatch
        - skillLevelMatch
        - goalTypeMatch

    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        studyTime:
          type: string
        goalType:
          type: string
        minSkillLevel:
          type: string
        requiredRoles:
          type: array
          items:
            $ref: "#/components/schemas/RoleRequirement"
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
        createdAt:
          type: string
      required:
        - id
        - name
        - subject
        - studyTime
        - goalType
        - minSkillLevel
        - requiredRoles
        - members
        - createdAt

    CreateTeamBody:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        studyTime:
          type: string
        goalType:
          type: string
        minSkillLevel:
          type: string
        requiredRoles:
          type: array
          items:
            $ref: "#/components/schemas/CreateRoleRequirement"
      required:
        - name
        - subject
        - studyTime
        - goalType
        - minSkillLevel
        - requiredRoles

    RoleRequirement:
      type: object
      properties:
        id:
          type: integer
        skillCategory:
          type: string
        count:
          type: integer
        filled:
          type: integer
      required:
        - id
        - skillCategory
        - count
        - filled

    CreateRoleRequirement:
      type: object
      properties:
        skillCategory:
          type: string
        count:
          type: integer
      required:
        - skillCategory
        - count

    TeamMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        teamId:
          type: integer
        role:
          type: string
        user:
          $ref: "#/components/schemas/User"
        joinedAt:
          type: string
      required:
        - id
        - userId
        - teamId
        - role
        - user
        - joinedAt

    AddTeamMemberBody:
      type: object
      properties:
        userId:
          type: integer
        role:
          type: string
      required:
        - userId
        - role

    DashboardStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalTeams:
          type: integer
        bySubject:
          type: array
          items:
            $ref: "#/components/schemas/SubjectCount"
        bySkillLevel:
          type: array
          items:
            $ref: "#/components/schemas/SkillLevelCount"
        byStudyTime:
          type: array
          items:
            $ref: "#/components/schemas/StudyTimeCount"
      required:
        - totalUsers
        - totalTeams
        - bySubject
        - bySkillLevel
        - byStudyTime

    SubjectCount:
      type: object
      properties:
        subject:
          type: string
        count:
          type: integer
      required:
        - subject
        - count

    SkillLevelCount:
      type: object
      properties:
        level:
          type: string
        count:
          type: integer
      required:
        - level
        - count

    StudyTimeCount:
      type: object
      properties:
        time:
          type: string
        count:
          type: integer
      required:
        - time
        - count